// packages/core/src/common/types.ts
export interface ColumnStatsData {
  column: string;
  nullCount: number;
  nullPct: number;
  distinctCount: number;
  min: any;
  max: any;
  avgLength?: number;
  totalRows: number;
}